+++ src/model/elements/ExitPoint.cpp
@@ -1,15 +1,14 @@
 #include "ExitPoint.hpp"
 
 #include "actions/ModelActionFactory.hpp"
+#include "actions/ModelActionUtils.hpp"
 #include "private/IModelVisitor.hpp"
 
 namespace model {
 
 ExitPoint::ExitPoint(const QString& name)
     : State(name, StateType::EXITPOINT)
-    , mOnStateChangedAction(ModelActionFactory::createModelAction(ModelAction::NONE))
-    , mOnEnteringAction(ModelActionFactory::createModelAction(ModelAction::NONE))
-    , mOnExitingAction(ModelActionFactory::createModelAction(ModelAction::NONE)) {}
+    , mOnStateChangedAction(ModelActionFactory::createModelAction(ModelAction::NONE)) {}
 
 void ExitPoint::accept(class IModelVisitor* visitor) {
     if (visitor) {
@@ -25,12 +24,21 @@ QSharedPointer<IModelAction> ExitPoint::onStateChangedAction() const {
     return mOnStateChangedAction;
 }
 
+const ModelActionList& ExitPoint::onEnteringActions() const {
+    return mOnEnteringActions;
+}
+
+const ModelActionList& ExitPoint::onExitingActions() const {
+    return mOnExitingActions;
+}
+
 QSharedPointer<IModelAction> ExitPoint::onEnteringAction() const {
-    return mOnEnteringAction;
+    return (mOnEnteringActions.isEmpty() ? ModelActionFactory::createModelAction(ModelAction::NONE)
+                                         : mOnEnteringActions.first());
 }
 
 QSharedPointer<IModelAction> ExitPoint::onExitingAction() const {
-    return mOnExitingAction;
+    return (mOnExitingActions.isEmpty() ? ModelActionFactory::createModelAction(ModelAction::NONE) : mOnExitingActions.first());
 }
 
 bool ExitPoint::hasOnStateChangedAction() const {
@@ -38,11 +46,11 @@ bool ExitPoint::hasOnStateChangedAction() const {
 }
 
 bool ExitPoint::hasOnEnteringAction() const {
-    return (mOnEnteringAction && mOnEnteringAction->type() != ModelAction::NONE);
+    return (mOnEnteringActions.isEmpty() == false);
 }
 
 bool ExitPoint::hasOnExitingAction() const {
-    return (mOnExitingAction && mOnExitingAction->type() != ModelAction::NONE);
+    return (mOnExitingActions.isEmpty() == false);
 }
 
 void ExitPoint::setEvent(const QString& event) {
@@ -55,18 +63,56 @@ void ExitPoint::setOnStateChangedAction(const QSharedPointer<IModelAction>& acti
     emit modelDataChanged(sharedFromThis().toWeakRef());
 }
 
-void ExitPoint::setOnEnteringAction(const QSharedPointer<IModelAction>& action) {
-    mOnEnteringAction = (action ? action : ModelActionFactory::createModelAction(ModelAction::NONE));
-    emit modelDataChanged(sharedFromThis().toWeakRef());
+void ExitPoint::setOnEnteringActions(const ModelActionList& actions) {
+    setActions(mOnEnteringActions, actions);
 }
 
-void ExitPoint::setOnExitingAction(const QSharedPointer<IModelAction>& action) {
-    mOnExitingAction = (action ? action : ModelActionFactory::createModelAction(ModelAction::NONE));
-    emit modelDataChanged(sharedFromThis().toWeakRef());
+void ExitPoint::setOnExitingActions(const ModelActionList& actions) {
+    setActions(mOnExitingActions, actions);
+}
+
+void ExitPoint::addOnEnteringAction(const QSharedPointer<IModelAction>& action) {
+    addAction(mOnEnteringActions, action);
+}
+
+void ExitPoint::addOnExitingAction(const QSharedPointer<IModelAction>& action) {
+    addAction(mOnExitingActions, action);
+}
+
+void ExitPoint::insertOnEnteringAction(const int index, const QSharedPointer<IModelAction>& action) {
+    insertAction(mOnEnteringActions, index, action);
+}
+
+void ExitPoint::insertOnExitingAction(const int index, const QSharedPointer<IModelAction>& action) {
+    insertAction(mOnExitingActions, index, action);
+}
+
+void ExitPoint::removeOnEnteringAction(const int index) {
+    removeAction(mOnEnteringActions, index);
+}
+
+void ExitPoint::removeOnExitingAction(const int index) {
+    removeAction(mOnExitingActions, index);
+}
+
+void ExitPoint::moveOnEnteringAction(const int from, const int to) {
+    moveAction(mOnEnteringActions, from, to);
+}
+
+void ExitPoint::moveOnExitingAction(const int from, const int to) {
+    moveAction(mOnExitingActions, from, to);
 }
 
 void ExitPoint::setOnStateChangedAction(const QString& actionData) {
     setOnStateChangedAction(ModelActionFactory::createModelActionFromData(actionData, ModelAction::CALLBACK));
+}
+
+void ExitPoint::setOnEnteringAction(const QSharedPointer<IModelAction>& action) {
+    setActions(mOnEnteringActions, ModelActionUtils::singleActionList(action));
+}
+
+void ExitPoint::setOnExitingAction(const QSharedPointer<IModelAction>& action) {
+    setActions(mOnExitingActions, ModelActionUtils::singleActionList(action));
 }
 
 void ExitPoint::setOnEnteringAction(const QString& actionData) {
@@ -93,13 +139,17 @@ bool ExitPoint::setProperty(const QString& key, const QVariant& value) {
             setOnStateChangedAction(value.toString());
         }
     } else if (key == "onEnteringAction") {
-        if (value.canConvert<QSharedPointer<IModelAction>>()) {
+        if (value.canConvert<ModelActionList>()) {
+            setOnEnteringActions(value.value<ModelActionList>());
+        } else if (value.canConvert<QSharedPointer<IModelAction>>()) {
             setOnEnteringAction(value.value<QSharedPointer<IModelAction>>());
         } else {
             setOnEnteringAction(value.toString());
         }
     } else if (key == "onExitingAction") {
-        if (value.canConvert<QSharedPointer<IModelAction>>()) {
+        if (value.canConvert<ModelActionList>()) {
+            setOnExitingActions(value.value<ModelActionList>());
+        } else if (value.canConvert<QSharedPointer<IModelAction>>()) {
             setOnExitingAction(value.value<QSharedPointer<IModelAction>>());
         } else {
             setOnExitingAction(value.toString());
@@ -112,17 +162,21 @@ bool ExitPoint::setProperty(const QString& key, const QVariant& value) {
 }
 
 QVariant ExitPoint::getProperty(const QString& key) const {
+    QVariant res;
+
     if (key == "event") {
-        return mEvent;
+        res = mEvent;
     } else if (key == "onStateChangedAction") {
-        return QVariant::fromValue(mOnStateChangedAction);
+        res = QVariant::fromValue(mOnStateChangedAction);
     } else if (key == "onEnteringAction") {
-        return QVariant::fromValue(mOnEnteringAction);
+        res = QVariant::fromValue(mOnEnteringActions);
     } else if (key == "onExitingAction") {
-        return QVariant::fromValue(mOnExitingAction);
+        res = QVariant::fromValue(mOnExitingActions);
+    } else {
+        res = State::getProperty(key);
     }
 
-    return State::getProperty(key);
+    return res;
 }
 
 void ExitPoint::copyEntityData(const StateMachineEntity& other) {
@@ -132,10 +186,8 @@ void ExitPoint::copyEntityData(const StateMachineEntity& other) {
         mEvent = eOther->mEvent;
         mOnStateChangedAction =
             ModelActionFactory::createModelActionFromData(eOther->onStateChangedAction()->serialize(), ModelAction::NONE);
-        mOnEnteringAction =
-            ModelActionFactory::createModelActionFromData(eOther->onEnteringAction()->serialize(), ModelAction::NONE);
-        mOnExitingAction =
-            ModelActionFactory::createModelActionFromData(eOther->onExitingAction()->serialize(), ModelAction::NONE);
+        mOnEnteringActions = ModelActionUtils::cloneActionList(eOther->onEnteringActions());
+        mOnExitingActions = ModelActionUtils::cloneActionList(eOther->onExitingActions());
     }
 }