+++ src/model/elements/RegularState.hpp
@@ -19,19 +19,36 @@ public:
 
     // Getters
     QSharedPointer<IModelAction> onStateChangedAction() const;
+    const ModelActionList& onEnteringActions() const;
+    const ModelActionList& onExitingActions() const;
+
+    // Backward-compatible single-action getters (first action, or a NONE action if empty)
     QSharedPointer<IModelAction> onEnteringAction() const;
     QSharedPointer<IModelAction> onExitingAction() const;
 
     // Setters
     void setOnStateChangedAction(const QSharedPointer<IModelAction>& action);
-    void setOnEnteringAction(const QSharedPointer<IModelAction>& action);
-    void setOnExitingAction(const QSharedPointer<IModelAction>& action);
+
+    // Multi-action list mutators (onEntering / onExiting)
+    void setOnEnteringActions(const ModelActionList& actions);
+    void setOnExitingActions(const ModelActionList& actions);
+    void addOnEnteringAction(const QSharedPointer<IModelAction>& action);
+    void addOnExitingAction(const QSharedPointer<IModelAction>& action);
+    void insertOnEnteringAction(const int index, const QSharedPointer<IModelAction>& action);
+    void insertOnExitingAction(const int index, const QSharedPointer<IModelAction>& action);
+    void removeOnEnteringAction(const int index);
+    void removeOnExitingAction(const int index);
+    void moveOnEnteringAction(const int from, const int to);
+    void moveOnExitingAction(const int from, const int to);
 
     bool hasOnStateChangedAction() const;
     bool hasOnEnteringAction() const;
     bool hasOnExitingAction() const;
 
     void setOnStateChangedAction(const QString& actionData);
+    // Backward-compatible single-action setters (reset the list to one action)
+    void setOnEnteringAction(const QSharedPointer<IModelAction>& action);
+    void setOnExitingAction(const QSharedPointer<IModelAction>& action);
     void setOnEnteringAction(const QString& actionData);
     void setOnExitingAction(const QString& actionData);
 
@@ -71,8 +88,8 @@ private:
 
 protected:
     QSharedPointer<IModelAction> mOnStateChangedAction;
-    QSharedPointer<IModelAction> mOnEnteringAction;
-    QSharedPointer<IModelAction> mOnExitingAction;
+    ModelActionList mOnEnteringActions;
+    ModelActionList mOnExitingActions;
 
     QList<QSharedPointer<StateMachineEntity>> mChildren;
 };